Layer model structure finished

+++ internal/models/feeder.go
@@ -1,8 +1,8 @@
 package models
 
 type Feeder struct {
-	ID int `json:"id"`
-	HouseID int `json:"house_id"`
+	ID int64 `json:"id"`
+	HouseID int64 `json:"house_id"`
 	MacAddress string `json:"mac_address"`
 	Name string `json:"name"`
 	PetType string `json:"pet_type"`
